Changed All Instances of `ssize_t` to `int`

+++ tests/files/c_source/camel_case_empty.h
@@ -33,7 +33,6 @@
 
 #include <stdint.h>
 #include <stdbool.h>
-#include <unistd.h>
 
 #ifndef EINVAL
 #    define EINVAL 22
@@ -62,7 +61,7 @@ struct camel_case_empty_message1_t {
  *
  * @return Size of packed data, or negative error code.
  */
-ssize_t camel_case_empty_message1_pack(
+int camel_case_empty_message1_pack(
     uint8_t *dst_p,
     const struct camel_case_empty_message1_t *src_p,
     size_t size);
