+++ tests/files/c_source/no_signals.c
@@ -32,7 +32,7 @@
 
 #include "no_signals.h"
 
-ssize_t no_signals_message1_pack(
+int no_signals_message1_pack(
     uint8_t *dst_p,
     const struct no_signals_message1_t *src_p,
     size_t size)
@@ -64,7 +64,7 @@ int no_signals_message1_unpack(
     return (0);
 }
 
-ssize_t no_signals_message2_pack(
+int no_signals_message2_pack(
     uint8_t *dst_p,
     const struct no_signals_message2_t *src_p,
     size_t size)
